--- Convertir/cafeterias_20241029_185638.csv
@@ -1,28 +1,0 @@
-"horario_cierre_pm","nombre","capacidad_ordenes","admin","horario_apertura_pm","horario_apertura_am","institucion","is_active","id","logo","precio_para_llevar","horario_cierre_am"
-"22:30","Zamari Gourmet - Sede Zúñiga","1000","ADMIN.ZAMARI_ZUÑIGA_EIA","12:00","09:00","2024-07-10 12:48:53.563592Universidad EIA","True","7","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/Zamari Gourmet.jpg","1500","11:59"
-"14:00","Nativos - ECI ","1000","ADMIN.NATIVOS_ECI_BOGOTA","13:00","09:00","2024-07-10 14:01:43.128600Escuela Colombiana de Ingenierí­a Julio Garavito","True","8","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/Logo NATIVOS.jpg","0","08:00"
-"01:03","Dogger - UPB","1000","ADMIN.DOGGER_UPB","01:02","01:00","2024-07-10 12:51:20.391000Universidad UPB","True","10","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/WhatsApp Image 2024-02-08 at 22.53.25.jpeg","0","01:01"
-"18:00","Coco Melón - CES Poblado","1000","ADMIN.COCOMELON_CES","12:01","07:00","2024-07-10 12:53:20.806773Universidad CES","True","3","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/LOGO - JPG.jpg","1400","12:00"
-"19:00","Dulcinea - CES Poblado","1000","ADMIN.DULCINEA_CES","12:00","06:20","2024-07-10 12:53:20.806773Universidad CES","True","2","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/dulcinea.png","0","11:59"
-"22:00","BOTY","1000","ADMIN.BOTY_UMB","12:00","00:00","2024-07-10 12:55:00.881529Universidad Manuela Beltrán","True","21","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/Logo BOTY.jpg","0","11:59"
-"20:35","Pimientoz UPB","1000","ADMIN.PIMIENTOS_UPB","12:00","07:00","2024-07-10 12:51:20.391000Universidad UPB","True","14","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/LOGO-ABAJO-FOOTER.webp","1000","11:59"
-"15:30","El Buñuelo - CES","1000","ADMIN.ELBUÑUELO_CES","12:00","06:30","2024-07-10 12:53:20.806773Universidad CES","True","18","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/LOGO.jpg","0","11:59"
-"23:59","FUNDACIÓN PUTCHI","1000","ADMIN.PUTCHI","12:01","00:03","2024-07-10 14:06:26.920274COLOSAL SOFTWARE","True","25","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/putchi.png","15000","11:59"
-"19:30","Taco Factory - UPB","1000","ADMIN.TACO_FACTORY_UPB","12:00","11:30","2024-07-10 12:51:20.391000Universidad UPB","True","12","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/logo (1).jpg","0","11:59"
-"15:03","Aldea Nikkei - EAFIT","1000","ADMIN.AldeaNikkei_EAFIT","12:01","09:00","2024-07-10 12:49:55.096977Universidad EAFIT","True","9","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/logo.jpg","3600","12:00"
-"15:22","Zamari Gourmet - Sede Palmas","1000","ADMID.ZAMARI_EIA","13:00","09:00","2024-07-10 12:08:21.620187Universidad EIA","False","4","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/IMG-20240124-WA0007.jpg","1500","12:59"
-"20:30","Chip Station","250","ADMIN.STATION_EAFIT","12:00","08:00","2024-07-10 12:49:55.096977Universidad EAFIT","True","26","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/WhatsApp Image 2024-10-10 at 6.34.11 PM.jpeg","0","11:59"
-"14:00","ServiExpress","1000","ADMIN.EXPRESS_EIA","12:01","07:00","2024-07-10 12:08:21.620187Universidad EIA","True","13","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/d61f5dd7-1a9b-46b4-abf8-6fdce09723df.jpeg","1700","12:00"
-"01:03","Pizzotas","1000","ADMIN.PIZZOTAS_CES","01:02","01:00","2024-07-10 12:53:20.806773Universidad CES","True","16","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/Logo.jpg","0","01:01"
-"01:03","El Buñuelo - EIA","1000","ADMIN.ELBUÑUELO_EIA","01:02","01:00","2024-07-10 12:08:21.620187Universidad EIA","True","17","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/LOGO.jpg","0","01:01"
-"01:03","Dogger - EIA","1000","ADMIN.DOGGER_EIA","01:02","01:00","2024-07-10 12:08:21.620187Universidad EIA","True","6","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/WhatsApp Image 2024-01-24 at 20.54.03_89962712.jpg","0","01:01"
-"17:20","Montana","1000","ADMIN.SERVIEXPRESS_UPB","12:00","07:50","2024-07-10 12:51:20.391000Universidad UPB","True","20","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/WhatsApp Image 2024-06-04 at 16.23.27.jpeg","0","11:59"
-"23:59","COLOSAL PETS","1000","esteban.colosal","13:00","06:00","2024-07-10 14:06:26.920274COLOSAL SOFTWARE","False","19","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/placa_negra.png","0","13:00"
-"19:00","Feroz Helados de Yogurt - UPB","1000","ADMIN.FEROZ_UPB","12:01","10:00","2024-07-10 12:51:20.391000Universidad UPB","True","1","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/feroz.png","0","12:00"
-"17:00","De Lolita - EIA","1000","ADMIN.LOLITA_EIA","13:00","08:10","2024-07-10 12:08:21.620187Universidad EIA","True","0","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/de_lolita.png","0","12:00"
-"18:00","La Cafetería - Idiomas","1000","ADMIN.LaCafeteria_EAFIT","12:01","08:00","2024-07-10 12:49:55.096977Universidad EAFIT","True","5","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/la cafetria.jpg20240127180200","1700","12:00"
-"01:47","Nativos - EIA","1000","ADMIN.NATIVOS_EIA","01:46","01:44","2024-07-10 12:08:21.620187Universidad EIA","True","22","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/Nativos logo lejos.jpg20240715105905","0","01:45"
-"19:45","EINSTEIN","1000","ADMIN.Einstein_USC","12:00","08:00","2024-07-10 14:05:17.791844Universidad Santiago","True","24","","0","11:59"
-"16:35","Tradiciones Culinarias","1000","ADMIN.TRADICIONES_IUSH","12:00","08:00","2024-07-10 13:58:07.205716Universidad IUSH","True","23","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/Logo Tradiciones.jpg","1600","11:59"
-"23:45","Pimientoz","120","ADMIN.PIMIENTOS_EAFIT","12:00","08:32","2024-07-10 12:49:55.096977Universidad EAFIT","True","15","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/LOGO-ABAJO-FOOTER.webp","0","11:59"
-"20:00","Taco Factory - EAFIT","100","ADMIN.TACOFACTORY_EAFIT","12:00","11:10","2024-07-10 12:49:55.096977Universidad EAFIT","True","11","https://appu-bucket-s3.s3.amazonaws.com/cafeterias/logo (1).jpg","0","11:59"
